add comments to the updated files
