+++ src/charts/RowSummaryBox.js
@@ -122,15 +122,18 @@ class RowSummaryBox extends BaseChart {
         const removedIndex = currentParam.indexOf(column);
         const nextParam =
             impact?.nextParam ?? currentParam.filter((field) => field !== column);
+
         // The image key is stored by param index, so removing that field is not
         // safely prunable; the chart must be deleted instead.
         if (this.config.image && this.config.image.param === removedIndex) {
             return true;
         }
+        // Adjust the stored image-param index
         if (removedIndex !== -1 && this.config.image && this.config.image.param > removedIndex) {
             this.config.image.param -= 1;
         }
 
+        // Call onColumnRemoved and update the relevant settings
         super.onColumnRemoved(column, {
             ...impact,
             action: "update",
@@ -138,9 +141,13 @@ class RowSummaryBox extends BaseChart {
         });
 
         if (removedIndex === -1) {
+            // Shared cleanup may still have handled tooltip/color/background
+            // filter state even though there was no visible summary section to remove.
             return false;
         }
         if (this.img_data && this.config.image && Array.isArray(this.config.param)) {
+            // `img_data` is runtime-only derived state, so keep it aligned with
+            // the new image param index after pruning params.
             this.img_data.key_column = this.config.param[this.config.image.param];
         }
